pyproject.toml
[project]
name = "transcription"
version = "0.1.0"
description = "A tool to convert lectures into LaTeX notes."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nltk>=3.9.1",
    "openai>=1.82.1",
    "openai-whisper>=20240930",
    "python-dotenv>=1.1.0",
]
